[Refactor]: EditProfileOrganism 컴포 나눔, 프로필 디렉토리
